#275 refactoring to two main modules created - servant: the "old" application and they submodules - api: the new REST server api layer
