+++ modules/servant/song_loader/song_loader.py
@@ -6,10 +6,10 @@ from dacite import from_dict
 
 from config.config_data import ConfigData
 from common.definitions import PSARC_INFO_FILE_CACHE_DIR, TMP_DIR
-from modules.database.db_manager import DBManager
-from modules.song_loader.rs_playlist_data import RsPlaylist
-from modules.song_loader.song_data import SongData, ArtistTitle
-from modules.song_loader.song_loader_helper import playlist_does_not_changed, check_cdlc_archive_dir, \
+from modules.servant.database.db_manager import DBManager
+from modules.servant.song_loader.rs_playlist_data import RsPlaylist
+from modules.servant.song_loader.song_data import SongData, ArtistTitle
+from modules.servant.song_loader.song_loader_helper import playlist_does_not_changed, check_cdlc_archive_dir, \
     check_rocksmith_cdlc_dir, update_tags_in_song_data, is_official, log_new_songs_found
 from utils import file_utils, psarc_reader
 from utils.collection_utils import is_not_empty, is_empty, repr_in_multi_line
